+++ AEDA_FEUP/AEDA_FEUP/Rent.cpp
@@ -6,7 +6,7 @@ Utilities u2;
 Menu rentMenu;
 
 // ALUGUER
-Rent::Rent(int nif, string typeRent, string name2 ,string c, Date dataI, Date dataF, float p, int n)
+Rent::Rent(long nif, string typeRent, string name2 ,string c, Date dataI, Date dataF, float p, int n)
 {
 	city = c; dataBegin = dataI; dataEnd = dataF; numPeople = n; price = p; this->typeRent = typeRent; name = name2; this->nif = nif;
 }
@@ -30,14 +30,14 @@ bool Rent::isValid(Date d1, Date d2)
 
 // HOTEL
 
-Hotel::Hotel(int nif, string typeRent, string name, string cidade, Date dataI, Date dataF, string type, float preco, int numOcupantes) : Rent(nif, typeRent, name, cidade, dataI, dataF, preco, numOcupantes)
+Hotel::Hotel(long nif, string typeRent, string name, string cidade, Date dataI, Date dataF, string type, float preco, int numOcupantes) : Rent(nif, typeRent, name, cidade, dataI, dataF, preco, numOcupantes)
 {
 	nameHotel = name;
 	this->type = type;
 
 }
 
-Hotel Hotel::buildRent(int nif)
+Hotel Hotel::buildRent(long nif)
 {
 	string city;
 	int n;
@@ -166,12 +166,12 @@ Hotel Hotel::buildRent(int nif)
 
 // BED'N'BREAKFAST
 
-bedNbreakfast::bedNbreakfast(int nif, string typeRent, string name, string cidade, Date dataI, Date dataF, float preco, int numOcupantes) : Rent(nif, typeRent, name ,cidade, dataI, dataF, preco, numOcupantes)
+bedNbreakfast::bedNbreakfast(long nif, string typeRent, string name, string cidade, Date dataI, Date dataF, float preco, int numOcupantes) : Rent(nif, typeRent, name ,cidade, dataI, dataF, preco, numOcupantes)
 {
 	this->namebnb = name;
 }
 
-bedNbreakfast bedNbreakfast::buildRent(int nif)
+bedNbreakfast bedNbreakfast::buildRent(long nif)
 {
 	Utilities u;
 	string city;
@@ -276,13 +276,13 @@ bedNbreakfast bedNbreakfast::buildRent(int nif)
 
 // SHARED-HOUSE
 
-sharedHouse::sharedHouse(int nif, string typeRent, string name, string cidade, Date dataI, Date dataF, float preco, int numOcupantes) : Rent(nif, typeRent, name, cidade, dataI, dataF, preco, numOcupantes)
+sharedHouse::sharedHouse(long nif, string typeRent, string name, string cidade, Date dataI, Date dataF, float preco, int numOcupantes) : Rent(nif, typeRent, name, cidade, dataI, dataF, preco, numOcupantes)
 {
 	this->nameSH = name;
 }
 
 
-sharedHouse sharedHouse::buildRent(int nif)
+sharedHouse sharedHouse::buildRent(long nif)
 {
 	string city, name;
 	int n;
@@ -395,12 +395,12 @@ sharedHouse sharedHouse::buildRent(int nif)
 
 // FLAT
 
-flat::flat(int nif, string typeRent, string name, string cidade, Date dataI, Date dataF, float preco, int numOcupantes) : Rent(nif, typeRent, name ,cidade, dataI, dataF,preco, numOcupantes)
+flat::flat(long nif, string typeRent, string name, string cidade, Date dataI, Date dataF, float preco, int numOcupantes) : Rent(nif, typeRent, name ,cidade, dataI, dataF,preco, numOcupantes)
 {
 	this->nameFlat = name;
 }
 
-flat flat::buildRent(int nif)
+flat flat::buildRent(long nif)
 {
 	string city;
 	int n;
@@ -505,7 +505,7 @@ flat flat::buildRent(int nif)
 
 // APARTMENT
 
-apartment::apartment(int nif, string tipoRent, string name, string cidade, Date dataI, Date dataF, float price, int numOcupantes, int numrooms, bool kitchen, bool suite, bool livingroom) : Rent(nif, tipoRent, name, cidade, dataI, dataF, price, numOcupantes)
+apartment::apartment(long nif, string tipoRent, string name, string cidade, Date dataI, Date dataF, float price, int numOcupantes, int numrooms, bool kitchen, bool suite, bool livingroom) : Rent(nif, tipoRent, name, cidade, dataI, dataF, price, numOcupantes)
 {
 	nameApartment = name;
 	hasKitchen = kitchen;
@@ -513,7 +513,7 @@ apartment::apartment(int nif, string tipoRent, string name, string cidade, Date 
 	hasLivingRoom = livingroom;
 }
 
-apartment apartment::buildRent(int nif)
+apartment apartment::buildRent(long nif)
 {
 	string city;
 	int n;
